requirements.txt
amqp==1.4.9
anyjson==0.3.3
APScheduler==3.7.0
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.3.0.23
bs4==0.0.1
celery==3.1.26.post2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
dj-database-url==0.5.0
Django==3.1.5
django-celery==3.3.1
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.4.0
kombu==3.0.37
lxml==4.6.2
Markdown==3.3.4
numpy==1.19.5
pkg-resources==0.0.0
postgres==3.0.0
prompt-toolkit==3.0.14
psycopg2==2.8.6
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pytz==2020.5
redis==2.10.3
requests==2.25.1
selenium==3.141.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
webdriver-manager==3.3.0
whitenoise==5.2.0
zipp==3.4.0
